feat/gas-refund: clear type system
