+++ scripts/gas-refund-program/types.ts
@@ -1,37 +1,6 @@
+import { PendingEpochGasRefundData } from "../../src/lib/gas-refund";
+
 export type HistoricalPrice = { timestamp: number; rate: number }[];
-
-interface BaseGasRefundData {
-  epoch: number;
-  address: string;
-  chainId: number;
-}
-export interface PendingEpochGasRefundData extends BaseGasRefundData {
-  accumulatedGasUsedPSP: string;
-  accumulatedGasUsed: string;
-  accumulatedGasUsedChainCurrency: string;
-  lastBlockNum: number;
-  isCompleted: false;
-  totalStakeAmountPSP: string;
-  refundedAmountPSP: string;
-  updated?: boolean;
-}
-
-export interface CompletedEpochGasRefundData
-  extends Partial<Omit<PendingEpochGasRefundData, 'isCompleted'>> {
-  merkleProofs: string[];
-  isCompleted: true;
-}
-
-export type EpochGasRefundData =
-  | PendingEpochGasRefundData
-  | CompletedEpochGasRefundData;
-
-export type GasRefundProgramdata = {
-  epoch: number;
-  chainId: number;
-  totalPSPAmountToRefund: string;
-  merkleRoot: string;
-};
 
 export type TxFeesByAddress = {
   [address: string]: PendingEpochGasRefundData;
@@ -62,9 +31,6 @@ export type MerkleTreeData = {
   leaves: MerkleData[];
 };
 
-export type MerkleTreeDataByChain = {
-  [chainId: number]: MerkleTreeData | null;
-};
 
 export type StakedPSPByAddress = {
   [address: string]: string;
