+++ tsconfig.json
@@ -17,5 +17,5 @@
     "emitDecoratorMetadata": true,
     "typeRoots": ["node_modules/@types"]
   },
-  "include": ["src/**/*", "scripts/gas-refund-program/**/*"]
+  "include": ["src/**/*"]
 }
